Added tests for the base item class
