+++ lib/noun-project-api/icon.rb
@@ -1,24 +1,15 @@
+require "noun-project-api/base_item"
+
 module NounProjectApi
   # A single Icon as an abstracted ruby object.
-  class Icon
+  class Icon < BaseItem
     PREVIEW_SIZE_200 = 200
     PREVIEW_SIZE_42 = 42
     PREVIEW_SIZE_84 = 84
 
     PUBLIC_DOMAIN_LICENSE = "public-domain".freeze
 
-    attr_accessor :original_hash
-
-    def initialize(origin)
-      origin = JSON.parse(origin) if origin.is_a? String
-      origin = origin.delete("icon") if origin.key? "icon"
-
-      @original_hash = origin
-    end
-
-    def id
-      original_hash["id"].to_i
-    end
+    ITEM_NAME = "icon".freeze
 
     def public_domain?
       original_hash["license_description"] == PUBLIC_DOMAIN_LICENSE
@@ -43,10 +34,6 @@ module NounProjectApi
         preview_url_84: preview_url(PREVIEW_SIZE_84),
         preview_url_42: preview_url(PREVIEW_SIZE_42)
       }
-    end
-
-    def to_json
-      JSON.dump(to_hash)
     end
   end
 end
