+++ spec/lib/noun-project-api/base_item_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+require "ostruct"
+
+RSpec.describe NounProjectApi::BaseItem do
+  it "raises an error on direct initialization" do
+    data = JSON.parse(Fakes::Results::ICON_VALID)
+    expect { NounProjectApi::BaseItem.new(JSON.dump(data["icon"])) }.to raise_error(NotImplementedError)
+  end
+end
